requirements-locked.txt
# DinoAir Complete Locked Requirements
# Generated on September 21, 2025
# All packages from all requirements files with exact versions
# Successfully installed and tested

aiofiles==24.1.0
alabaster==1.0.0
annotated-types==0.7.0
anyio==4.10.0
Authlib==1.6.4
babel==2.17.0
bandit==1.8.6
black==24.3.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
coverage==7.10.7
cryptography==46.0.1
docutils==0.22.2
dparse==0.6.4
fastapi==0.117.1
filelock==3.19.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
imagesize==1.4.1
iniconfig==2.1.0
isort==6.0.1
Jinja2==3.1.6
joblib==1.5.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==4.0.1
mdurl==0.1.2
memory-profiler==0.61.0
mypy==1.18.2
mypy_extensions==1.1.0
nltk==3.9.1
numpy==2.3.3
orjson==3.9.15
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
psutil==7.1.0
py-spy==0.4.1
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pypdf==6.1.0
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==7.0.0
pytest-mock==3.15.1
pytest-timeout==2.4.0
python-dotenv==1.1.1
python-json-logger==3.3.0
PyYAML==6.0.2
regex==2025.9.18
requests==2.32.5
rich==14.1.0
roman-numerals-py==3.1.0
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
ruff==0.13.0
safety==3.6.2
safety-schemas==0.0.16
shellingham==1.5.4
sniffio==1.3.1
snowballstemmer==3.0.1
Sphinx==8.2.3
sphinx-autodoc-typehints==3.2.0
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
starlette==0.48.0
stevedore==5.5.0
tenacity==9.1.2
tomlkit==0.13.3
tqdm==4.67.1
typer==0.19.1
types-aiofiles==24.1.0.20250822
types-psutil==5.9.5.20240516
types-requests==2.32.4.20250913
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.36.0
watchfiles==1.1.0
websockets==15.0.1
